requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
python-multipart==0.0.20
pydantic==2.11.3
python-dotenv==1.1.0
pypdf==5.4.0
chromadb==1.0.7
google-genai==1.12.1
streamlit==1.44.1
requests==2.32.3
pytest==8.3.5
httpx==0.28.1